IPFS: Factor out GatewayURL
