+++ protocol/ipfs/gatewayurl_test.go
@@ -1,8 +1,9 @@
 package ipfs
 
 import (
-	"github.com/stretchr/testify/suite"
 	"net/http"
+
+	"github.com/stretchr/testify/suite"
 	"testing"
 
 	"github.com/ipfs-search/ipfs-search/instr"
@@ -11,20 +12,20 @@ import (
 
 const gatewayURL = "http://ipfs:8080"
 
-type IPFSTestSuite struct {
+type GatewayURLTestSuite struct {
 	suite.Suite
 
 	ipfs *IPFS
 }
 
-func (s *IPFSTestSuite) SetupTest() {
+func (s *GatewayURLTestSuite) SetupTest() {
 	cfg := DefaultConfig()
 	cfg.GatewayURL = gatewayURL
 
 	s.ipfs = New(cfg, http.DefaultClient, instr.New())
 }
 
-func (s *IPFSTestSuite) TestGatewayURLReferenced() {
+func (s *GatewayURLTestSuite) TestGatewayURLReferenced() {
 	r := &t.AnnotatedResource{
 		Resource: &t.Resource{
 			Protocol: t.IPFSProtocol,
@@ -44,7 +45,7 @@ func (s *IPFSTestSuite) TestGatewayURLReferenced() {
 	s.Equal(url, gatewayURL+"/ipfs/QmSKboVigcD3AY4kLsob117KJcMHvMUu6vNFqk1PQzYUpp/fileName.pdf")
 }
 
-func (s *IPFSTestSuite) TestGatewayURLUnreferenced() {
+func (s *GatewayURLTestSuite) TestGatewayURLUnreferenced() {
 	r := &t.AnnotatedResource{
 		Resource: &t.Resource{
 			Protocol: t.IPFSProtocol,
@@ -57,7 +58,7 @@ func (s *IPFSTestSuite) TestGatewayURLUnreferenced() {
 	s.Equal(url, gatewayURL+"/ipfs/QmcBLKyRHjbGeLnjnmj74FFJpGJDz4YxFqUDYqMU7Mny1p")
 }
 
-func (s *IPFSTestSuite) TestGatewayURLUnnamedReference() {
+func (s *GatewayURLTestSuite) TestGatewayURLUnnamedReference() {
 	r := &t.AnnotatedResource{
 		Resource: &t.Resource{
 			Protocol: t.IPFSProtocol,
@@ -77,6 +78,6 @@ func (s *IPFSTestSuite) TestGatewayURLUnnamedReference() {
 	s.Equal(url, gatewayURL+"/ipfs/QmcBLKyRHjbGeLnjnmj74FFJpGJDz4YxFqUDYqMU7Mny1p")
 }
 
-func TestIPFSTestSuite(t *testing.T) {
-	suite.Run(t, new(IPFSTestSuite))
+func TestGatewayURLTestSuite(t *testing.T) {
+	suite.Run(t, new(GatewayURLTestSuite))
 }
